修复一些警告

+++ src/pump_light/struct_reflection.hpp
@@ -45,14 +45,14 @@ struct pump_light_detect_param {
     struct optional_params {
         struct dyparams_params {
             float conf_thres;
-            float x1;
-            float y1;
-            float x2;
-            float y2;
-            float x3;
-            float y3;
-            float x4;
-            float y4;
+            int x1;
+            int y1;
+            int x2;
+            int y2;
+            int x3;
+            int y3;
+            int x4;
+            int y4;
         };
         dyparams_params dyparams;
     };
